Move monitoring router parameters into object attributes
This is to support rearrangement of the structure of the router code into
multiple threads and methods, without having to manually wire all of the
multiprocessing objects between the new methods and threads.

These objects are part of the context of the router object, rather than
parameters to individual methods which might change, and they are all
multiprocessing objects which are thread-safe.