Make WorkQueue scaling aware of core counts (#3415)
Prior to this PR, the scaling behaviour assumed that each worker could execute one task. That was/is true when no `parsl_resource_specification` is supplied, but results in over-scaling when a core count is supplied.

This PR pays attention to core count when specified and allows the user to describe how many cores a worker will be assumed to have for scaling purposes.