Terminate htex result queue thread by polling for an exit flag
Before this PR, this thread stays running forever

this *requires* socket to be closed at exit -- and this PR introduces code to do that

context:
see recent flux PR for same problem.
because stopping this thread is now allowing garbage collection to happen, it looks like? or something similar... see PR #3517 for the same problem in Flux


counts:
before this PR, on parsl/tests/test_monitoring/

451 fds, 32 threads

after this PR, 48 fds, 1 thread