Treat python_app input/output default arguments like other magic keywords, and like bash_app (#3489)
In a previous PR (#3485), some fields where the default values of outputs and inputs for AppBase were removed. These fields were not being used in any way. However, like the other parsl reserved parameters, the defaults should have been stored in the AppBase.kwargs dict which is used to ensure that the default values of these special reserved parameters are in invocation_kwargs which is sent to the dataflow kernel and used to resolve the corresponding special behaviors. The correct behavior was only observed in BashApp since it does additional checking to store all default args (even the non-reserved ones) to invocation_kwargs. Now it should be consistent for both types of apps.

Changed Behaviour:
PythonApp will now correctly resolve default arguments for the inputs and outputs special parameters.