Move FluxExecutor ZMQ into thread and explicitly clean it up
Prior to this PR, there were frequent hangs in CI at cleanup of the
ZMQ objects used by the FluxExecutor. See issue #3484 for some more
information.

This PR attempts to remove some dangerous behaviour there:

i) creation of ZMQ context and socket is moved into the thread which
makes use of them - previous the socket was created on the main thread
and passed into the submission thread which uses it. This removes some
thread safety issues where a socket cannot be safely moved between
threads.

ii) ZMQ context and socket are more explicitly closed (using with-blocks)
rather than leaving that to the garbage collector. In the hung tests,
the ZMQ context was being garbage collected in the main thread, which is
documented as being unsafe when sockets are open belonging to another
thread (the submission thread)

On my laptop I could see a hang around 50% of test runs before this PR.
After this PR, I have run about 100 iterations of the flux tests with seeing
any hangs.