Close Channels at DFK cleanup (#3503)
Following on from #3403, which tightens the type signature of unused Channel.close(), this PR makes the DFK actually call Channel.close() when it shuts down an executor.

At startup, the DFK chooses whether to manage things with channels by checking for the presence of a script_dir attribute on the relevant provider (!) and this PR sticks with that same informal protocol.

On the LocalChannel, this has no effect. On the SSH Channels, this shuts down the SSH connection and removes an otherwise-abandoned thread.