Use caplog fixure rather than mock logging when testing HTEX shutdown
This replaces mock-captured call sequences which are over-testing the exact
position of tested log calls in relation to any other log calls, which makes
this test fragile when working on logs in the affected area of the code.

This is consistent with other parts of the test suite which test log messages
using caplog.