Refactor naive scale in behaviour for Work Queue and Task Vine
The intended behaviour of this scale in code, which is only for scaling in
all blocks (for example at the end of a workflow) makes sense as a default
for all BlockProviderExecutors. This PR makes that refactor.

This code is buggy (before and after) - see issue #3471. This PR does not
attempt to fix that, but moves code into a better place for bugfixing, and
a subsequent PR will fix it.