the job submission PENDING status update has this principle so that scaling strategy will see submitted jobs immediately, before a provider status refresh happens. this makes the scaling code immediately aware of what just happened, rather than for one poll period acting as if nothing had happened.
When making changes that will later be reflected in the _status table, then those changes should be immediately also be made in the cached _status table eplciitly.

before this PR, this code path does not happen in the case of a failed submission, where a failure status will appear when a refresh happens, but not before. in that case the scaling code will act as if the failed submission did not happen and will continue to submit repeatedly until a refresh happens much later.  this PR makes _status be updated in this case too.

fixes 3235