Remove monitoring router modification of node message
Prior to this PR, the monitoring router would add a run_id field to every
NODE_INFO message that it received. These are messages from the interchange
describing worker pools.

The monitoring router does not modify any other messages.

This PR sets the run_id at the point of message origination inside the
interchange (in _send_monitoring_info), and makes the router leave NODE_INFO
messages unmodified (like the other message types).

This is part of work to make the router less aware of message types by
removing a bunch of message-type specific handling.

This PR brings in a bunch of rewiring to get the run id into the interchange
rather than into the monitoring router.