the aim here is to move all checkpointing and memoizationc alls into the memoizer in anticipation of that class becoming pluggable.
the DFK tells the memoizer that things have happened but doesn't drive any
specific behaviour. that's the behaviour that would become pluggable.

because it's going to become pluggable, the Memoizer object should be separated into
creation (currently in the DFK but ultimately by the user) and .start(), triggered
by the DFK, and like other create vs deferred initialization config objects in
Parsl.

TODO: this should get a close method for the memoizer to trigger the cleanup behaviour
which is then later abstracted into an API close method