configurable memoizer instance
TODO: rewrite documentation that refers to any example checkpointing or app caching or memoization configuration
with the documentation ready to receive documentation for the SQL plugin

this PR is the biggest user facing change because it changes the configuration interface for anyone configuring
anything to do with app caching/memoization/checkpointing.

TODO: any tests that refer to config.checkpoint_whatever - they aren't getting detected as now testing the
wrong thing by mypy/flake8.