Remove explicit pytest flux start, because FluxExecutor does the real flux start (#3511)
Prior to this PR, flux pytests were run inside a flux start command; but inside that, FluxExecutor does its owns flux start - see around line 177 in  parsl/parsl/executors/flux/executor.py

This second, inner flux is what is used to execute tasks as if it was a batch allocation on a cluster.

So the outer pytest flux is only used to run the coordinating test workflow and launch that inner flux, as if it were on a submitting/login node. This is unnecessary.