Only scale in blocks that are in _status and non-terminal (#3548)
In the BlockProviderExecutor, the block ID to job ID mapping structures contain the full historical list of blocks. Prior to this PR, the mapping was used as source of current jobs that should/could be scaled in.

This was incorrect. and resulted in scaling in code attempting to:

scale in blocks that had already finished, because it continues to see those blocks as eligible for scale-in

not scale in blocks that were active - because rather than choosing to scale in an alive block, the code would choose to attempt to scale in a non-alive block

After this PR, the _status structure which should contain reasonably up to date status information is used instead of the block/job ID mapping structures. (as a more general principle, those block/job ID mapping structures should never be examined as a whole but only used for mapping)

Changed Behaviour: 
Scaling in should work better in executors using the default scaling in that was refactored in PR #3526, which right now is Work Queue and Task Vine.

Fixes #3471