Aggressively deprecate Channels and AdHocProvider
This is described in issue #3515.

Issue #3515 has not received any comments arguing in favour of retaining
channels and the AdHocprovider, and several in support of removing them,
 and so this PR takes a heavy handed approach that is well on the way to
the end goal of #3515 of deleting channels and the AdHocProvider entirely:

Channels except LocalChannel are renamed, so that any users of
other channels will have to make a change to their code and
actively observe the word "Deprecated" in the name.

The AdHocProvider is renamed in the same way.

Most documentation (but not docstrings) about channels and the
ad-hoc provider is removed or replaced with a link to issue

Tests which much be manually run, and so in effect are never run and
shouldn't be expected to work now - parsl/tests/manual_tests and
parsl/tests/integration/ - are deleted rather than fixed to follow
the above naming change.

The tests for SSH channels and the AdHocProvider that run in CI are modified
to continue passing.

Exposure of the deprecated components via top level parsl.providers and
parsl.channels re-export is removed. To use this components, the deprecated
modules must be imported directly.