Add an out-of-memory sqlite backed memoizer
This memoizer does not make a distinction between memoization and
checkpointing (which the default BasicMemoizer does: memoization is the
first stage in memory, and checkpointing is later on disk).

Instead, all operations about whether a task is memoized are presented
to an sqlite database. This can make lookup operations slower, but the
tradeoff is that no memory is occupied by the memoization table. The
main usecase for this is checkpoint restore in situations where a lot
of memory is used by the default memoizer's restore of all tasks every
run.

The BasicMemoizer handles exceptions differently between memoization and
checkpointing: exceptions are memoized (within a run) but not checkpointed.

As this new sqlite memoizer does not have that distinction, this code
makes an arbitrary choice to not record exceptions, even in memory.

The code is open to implementing exception memoization/checkpointing
in a future PR at the same time as it would be implemented for BasicMemoizer.