Remove manual checkpoint() directory return value
Prior to this PR, a manual checkpoint() call returned the directory that the
checkpoint was written to.

This isn't really needed, so this PR removes it as part of removing cruft
around the checkpoint API, as part of refactoring towards pluggable
checkpointing.

The only use is in a test case, which this PR rewrites.