Add breadcrumbs for users trying to match configs to their machines (#3603)
Both Polaris and Perlmutter have configuration documentation hosted by their respective facilities. This is great, but it makes it harder for users trying to match an undocumented machine to existing configs. This PR adds some hints to both Polaris and Perlmutter listing the provider and launcher combo used in the hopes that it might lead the user to configs from those machines.