Unhide pytest output by removing dotfile directory
Prior to this PR, output from pytest was generally hidden
by placing it in the .pytest directory.

After this PR, that .pytest directory is now pytest-parsl.

Most immediately, a breaking change in the GitHub actions
artifact upload has started respecting that dotfile
hiddenness, and so with that version of the artifact uploader,
test runs prior to this PR are missing a bunch of output.
https://github.com/actions/upload-artifact/issues/602

The uploader now has an option to enable hidden files...

but ...

.. before this most immediate motivation, I've run into
a couple of developer experience cases that look like:

  Q: I cannot find the output of pytest
  A: We have deliberately made it hard for you to find

which is enough to make me rename the directory rather than
turn on the artifact uploader hidden files option.