Remove interchange multiprocess log hack (#3644)
Prior to PR #3463, the interchange process was launched
with multiprocessing fork and inherited the log configuration
of the parent process.


To give the interchange its own log file, a specific logger,
called "interchange" was used. In PR #2307, this logger was
configured to not propagate entries upwards, so that user
defined root handlers in the parent process do not see
interchange logs.

Since #3463 this special configuration has no longer been
necessary. See #3635 for a related change to signal handlers.

Co-authored-by: Kevin Hunter Kesling <kevin@globus.org>