Combine provider and channel script dir initialization and creation
This is on the way to removing channel script dir entirely

The channel script dir will not differ from the provider script dir
any more - they were different because one was on the remote file
system and one was on the local filesystem.

Future code modifications can be clearer that these two are the same.