Rework environment handling in LocalChannel
Environments are never passed to local channel, so the environment parameter
is removed from execute_wait by this PR.

This will change the environment behaviour of the local channel: commands
will now receive a copy of the parent environment at time of command
invocation. Prior to this PR, commands receive a copy of the parent
environment at the time of channel initialization and would not see any
changes made to the process environment after that.

This new behaviour is more consistent with Popen. It is a user-facing
change, though.

Because it is user facing and not directly part of #3515 channel removal,
it could be made into a PR now.