Add resilience for heartbeats from unknown managers (#3643)
# Description

Make interchange survive a heartbeat message from an unregistered
manager.

#3262 and #3632 report situations where a heartbeat is received by the
interchange after the heartbeat period has expired, and so the relevant
manager has been unregistered. Before this PR, the interchange crashed
when this happened. After this PR, it will log a warning.

This PR adds a test for such a heartbeat message as well as a couple of
other bad protocol messages.

# Fixes

Fixes #3262, fixes #3632

## Type of change

- Bug fix