Be explicit about filesystem monitoring radio being ForkProcess
Prior to this PR, on Linux, this is already a fork process because of
defaults.

But this process *must* be a fork process, for the same reason that other
explicit ForkProcesses in Parsl must be fork processes: we cannot re-import
the the unknown user workflow script again.

This PR makes that explicit and consistent with other forked processes.

Longer term, multiprocessing-forked processes need to go away throughout
Parsl - see issue #2343, but this PR is not for that.