Remove ability to set monitoring log directory separate from rundir
Prior to this PR, MonitoringHub had a logdir parameter which let the log
directory be set separately from the DFK-level run directory.

Other Parsl components generally don't let the user set this unless there is
a specific reason.

So this PR removes that feature, reducing the amount of state to be threaded
around.

When reading this patch, note that what the DFK calls the rundir is a different
directory vs what a Config object calls the rundir.