Add `shared_fs` and `staging_required` test markers (#3669)
# Description

This PR adds a new `staging_required` marker and a marker to several
tests that assume a shared filesystem to work. Similarly, a few tests
that use staging are now marked with the `staging_required` marker.

This PR splits out changes from #3607.

# Changed Behaviour

These markers should not affect any test-behavior since none of the test
entries in the Makefile make use of this.
These change will kick-in once they are used by `GlobusComputeExecutor`
and potentially `KubernetesProvider`
for tests.

## Type of change

Choose which options apply, and delete the ones which do not apply.

- New feature
- Code maintenance/cleanup