Be explicit about filesystem monitoring radio being ForkProcess (#3680)
Prior to this PR, on Linux, this is already a fork process because of
defaults.

But this process *must* be a fork process, for the same reason that
other explicit ForkProcesses in Parsl must be fork processes: we cannot
re-import the the unknown user workflow script again.

This PR makes that explicit and consistent with other forked processes.

Longer term, multiprocessing-forked processes need to go away throughout
Parsl - see issue #2343, but this PR is not for that.

# Changed Behaviour

This should not change behaviour on Linux. On platforms that are not
Linux, nothing is supported and is probably already broken here.

## Type of change

- Code maintenance/cleanup