Remove the opportunity to configure script_dir for LocalChannel
Prior to this PR, if the argument defaulted to None, then the DFK would
set a script dir based on the workflow rundir.

Post this PR, this default behaviour is the only behaviour and the user
cannot specify an overriding value.

Future PRs will remove channel level script directories entirely.
This is part of issues #3515 removing channels,