Remove channel push_file support from SlurmProvider
This has been a basically dead code path since removal of non-remote
channels:

Either a user chooses to not move files (overriding move_files to False)
and the push_file was not used.

Or the default push_file was used, but:
* this is always LocalChannel now (PR #3677)
* the script directory is always the local script directory (PR #3688)
and so the local channel code always skips making a copy

This commit simplies all of that away into using the script path directly
with putting it through the above complicated no-op.