Fix filesystem radio type annotation: the wrong kind of Queue (#3691)
Prior to this PR, filesystem_receiver was annotated as taking
queue.Queue. This was incorrect (but not detected or enforced by any
tooling - because type annotations are not carried across the invoking
Process constructor)

The receiver actually takes a multiprocessing.Queue, which is used
without a namespace, like the rest of this source file.

## Type of change

- Code maintenance/cleanup