Remove provider/channel push/pull of files (#3690)
This removes push/pull support from channels.

All the HPC providers used channel.push_file to push their batch scripts
to the (no-longer) remote system. This has been a basically dead code
path since removal of non-remote channels:

The channel is always a LocalChannel now (PR #3677)
The "remote" script directory is always the local script directory (PR
#3688) and so `LocalChannel.push_file` always skips making a copy and
returns the path it was given without further action.

So all the removed code is a no-op, and this PR simplifies that into
nothing.

# Changed Behaviour

Some providers had an option to let users decide if scripts (and other
files) would be pushed/pulled to the remote system. Those options are
removed by this PR.

## Type of change

- New feature
- Code maintenance/cleanup